🚚 move: gameEnd함수 파일 utils로 이동
